requirements.txt
flask
numpy
pandas
python-dotenv
requests
pymongo
OpenAI